Fix API generation
