+++ api/proto/teleport/devicetrust/v1/tpm.proto
@@ -57,41 +57,3 @@ message TPMPlatformAttestation {
   bytes nonce = 1;
   TPMPlatformParameters platform_parameters = 2;
 }
-
-// The attestation key and the parameters necessary to remotely verify it as
-// related to the endorsement key.
-// See https://pkg.go.dev/github.com/google/go-attestation/attest#AttestationParameters.
-// This message excludes the `UseTCSDActivationFormat` field from the link above
-// as it is TMP 1.x specific and always false.
-message TPMAttestationParameters {
-  // The encoded TPMT_PUBLIC structure containing the attestation public key
-  // and signing parameters.
-  bytes public = 1;
-  // The properties of the attestation key, encoded as a TPMS_CREATION_DATA
-  // structure.
-  bytes create_data = 2;
-  // An assertion as to the details of the key, encoded as a TPMS_ATTEST
-  // structure.
-  bytes create_attestation = 3;
-  // A signature of create_attestation, encoded as a TPMT_SIGNATURE structure.
-  bytes create_signature = 4;
-}
-
-// These values are used by the TPM2.0 `ActivateCredential` command to produce
-// the solution which proves possession of the EK and AK.
-//
-// For a more in-depth description see:
-// - https://pkg.go.dev/github.com/google/go-attestation/attest#EncryptedCredential
-// - https://trustedcomputinggroup.org/wp-content/uploads/TCG_TPM2_r1p59_Part3_Commands_code_pub.pdf (Heading 12.5.1 "TPM2_ActivateCredential" "General Description")
-// - https://github.com/google/go-attestation/blob/v0.4.3/attest/activation.go#L199
-// - https://github.com/google/go-tpm/blob/v0.3.3/tpm2/credactivation/credential_activation.go#L61
-message TPMEncryptedCredential {
-  // The `credential_blob` parameter to be used with the `ActivateCredential`
-  // command. This is used with the decrypted value of `secret` in a
-  // cryptographic process to decrypt the solution.
-  bytes credential_blob = 1;
-  // The `secret` parameter to be used with `ActivateCredential`. This is a
-  // seed which can be decrypted with the EK. The decrypted seed is then used
-  // when decrypting `credential_blob`.
-  bytes secret = 2;
-}
