+++ api/proto/teleport/legacy/types/types.proto
@@ -1273,14 +1273,36 @@ message ProvisionTokenSpecV2 {
 
 message ProvisionTokenSpecV2TPM {
   message Rule {
+    // Description is a human-readable description of the rule. It has no
+    // bearing on whether or not a TPM is allowed to join, but can be used
+    // to associate a rule with a specific host (e.g the asset tag of the server
+    // in which the TPM resides).
+    // Example: "build-server-100"
     string Description = 1 [(gogoproto.jsontag) = "description,omitempty"];
     oneof Identifier {
+      // EKPubHash is the SHA256 hash of the EKPub marshaled in PKIX format
+      // and encoded in hexadecimal. This value will also be checked when a TPM
+      // has submitted an EKCert, and the public key in the EKCert will be used
+      // for this check.
       string EKPubHash = 2 [(gogoproto.jsontag) = "ekpub_hash,omitempty"];
+      // EKCertSerial is the serial number of the EKCert in hexadecimal with
+      // colon separated nibbles. This value will not be checked when a TPM
+      // does not have an EKCert configured.
+      // Example: 73:df:dc:bd:af:ef:8a:d8:15:2e:96:71:7a:3e:7f:a4
       string EKCertSerial = 3 [(gogoproto.jsontag) = "ekcert_serial,omitempty"];
     }
   }
+  // Allow is a list of Rules, the presented delegated identity must match one
+  // allow rule to permit joining.
   repeated Rule Allow = 1 [(gogoproto.jsontag) = "allow,omitempty"];
-  repeated bytes EKCertAllowedCAs = 2 [(gogoproto.jsontag) = "ekcert_allowed_cas,omitempty"];
+  // EKCertAllowedCAs is a list of CA certificates that will be used to validate
+  // TPM EKCerts.
+  // When specified, joining TPMs must present an EKCert signed by one of the
+  // specified CAs. TPMs that do not present an EKCert will be not permitted to
+  // join.
+  // When unspecified, TPMs will be allowed to join with either an EKCert or an
+  // EKPubHash.
+  repeated string EKCertAllowedCAs = 2 [(gogoproto.jsontag) = "ekcert_allowed_cas,omitempty"];
 }
 
 // ProvisionTokenSpecV2Github contains the GitHub-specific part of the
